Основной main был разбит на 3 части для дальнейшего удобства в использовании тестов
